move settings related widgets to feature
