+++ lib/widgets/pages/home_page.dart
@@ -8,12 +8,12 @@ import 'package:coffeecard/features/leaderboard/presentation/pages/leaderboard_p
 import 'package:coffeecard/features/opening_hours/opening_hours.dart';
 import 'package:coffeecard/features/receipt/presentation/cubit/receipt_cubit.dart';
 import 'package:coffeecard/features/receipt/presentation/pages/receipts_page.dart';
+import 'package:coffeecard/features/settings/presentation/pages/settings_page.dart';
 import 'package:coffeecard/features/ticket/presentation/cubit/tickets_cubit.dart';
 import 'package:coffeecard/features/ticket/presentation/pages/tickets_page.dart';
 import 'package:coffeecard/features/user/presentation/cubit/user_cubit.dart';
 import 'package:coffeecard/service_locator.dart';
 import 'package:coffeecard/widgets/components/helpers/lazy_indexed_stack.dart';
-import 'package:coffeecard/widgets/pages/settings/settings_page.dart';
 import 'package:coffeecard/widgets/routers/app_flow.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
